+++ src/main/java/ru/epam/miniparking/service/OfficeService.java
@@ -1,28 +1,27 @@
 package ru.epam.miniparking.service;
 
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.epam.miniparking.domain.Office;
 import ru.epam.miniparking.exception.NotFoundException;
 import ru.epam.miniparking.repo.OfficeRepo;
 
+import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
 
 @Service
+@Transactional
+@AllArgsConstructor
 public class OfficeService {
     private final OfficeRepo officeRepo;
-
-    public OfficeService(OfficeRepo officeRepo) {
-        this.officeRepo = officeRepo;
-    }
 
     public Office save(Office newOffice) {
         return officeRepo.save(newOffice);
     }
 
     public Office findById(long id) {
-        return officeRepo.findById(id).orElseThrow(() ->
-                new NotFoundException(Office.class.getSimpleName(), id));
+        return officeRepo.findById(id).orElseThrow(() -> new NotFoundException(Office.class.getSimpleName(), id));
     }
 
     public List<Office> findByOfficeTitle(String title) {
@@ -40,13 +39,14 @@ public class OfficeService {
 
     public void deleteById(long id) {
         Optional<Office> o = officeRepo.findById(id);
+        if (!o.isPresent()) {
+            throw new NotFoundException(Office.class.getSimpleName(), id);
+        }
         o.ifPresent(office -> office.getDrivers().forEach(d -> {
             d.setSpot(null);
             d.setOffice(null);
+            officeRepo.deleteById(id);
         }));
-        o.orElseThrow(() -> new NotFoundException(Office.class.getSimpleName(), id));
-
-        officeRepo.deleteById(id);
     }
 
     public List<Office> findAll() {
